+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 
 ### 0.3.1
  - BeautifulSoupを入れたので、URLタグの実装
- - **[WIP] 0.3.1.1** defaultの挙動を修正する
+ - **0.3.1.1** defaultの挙動を修正する
    - 簡単に使えるように、引数をオプションにする
      - defaultはcurrent files(`.`)ディレクトリを監視するようにする
+ - **[WIP] 0.3.1.2**  Watch pathの引数がfile path(not directory path)だった場合に単独のファイルを監視するようにする
+ - **[WIP] 0.3.1.3**  yaml fileのHot reloadを行うかどうかをYaml configに書けるようにする
 
 ### 0.3.0
 - **0.3.0.2**URL Passの接続をちゃんとしたロジックで行う